separated blood crawler on wall and not on wall into two different npcs and prevented switching between them, not tested yet
